pyproject.toml
[project]
name = "image-mcp-gen"
version = "0.1.0"
description = "A lightweight, model-friendly MCP server that delivers random placeholder images."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.2",
    "pillow>=11.3.0",
    "requests>=2.32.4",
]
